nfe-modbus-energy-logger/requirements.txt
pymodbus>=3.0.0
pyyaml>=6.0
